src/api/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
mysql-connector-python==9.2.0
pandas==2.2.3
python-dotenv==1.0.1